requirements.txt
# main dependencies
django==4.2.13
gunicorn==22.0.0
psycopg2-binary==2.9.9
pillow==10.3.0
sorl-thumbnail==12.10.0
faker==25.9.2
requests==2.32.3
django-crispy-forms==2.2
crispy-bootstrap5==2024.2